@@ -18,20 +18,20 @@
     "directory": "package"
   },
   "dependencies": {
-    "@angular/animations": "9.1.7",
-    "@angular/cdk": "9.2.3",
-    "@angular/common": "9.1.7",
-    "@angular/core": "9.1.7",
-    "@angular/forms": "9.1.7",
-    "@angular/language-service": "9.1.7",
-    "@angular/platform-browser": "9.1.7",
-    "@angular/platform-browser-dynamic": "9.1.7",
-    "@angular/router": "9.1.7",
+    "@angular/animations": "10.1.4",
+    "@angular/cdk": "10.2.4",
+    "@angular/common": "10.1.4",
+    "@angular/core": "10.1.4",
+    "@angular/forms": "10.1.4",
+    "@angular/language-service": "10.1.4",
+    "@angular/platform-browser": "10.1.4",
+    "@angular/platform-browser-dynamic": "10.1.4",
+    "@angular/router": "10.1.4",
     "@biesbjerg/ngx-translate-extract-marker": "^1.0.0",
-    "@clr/angular": "^3.0.0",
+    "@clr/angular": "^4.0.3",
     "@clr/core": "^3.0.0",
-    "@clr/icons": "^3.0.0",
-    "@clr/ui": "^3.0.0",
+    "@clr/icons": "^4.0.3",
+    "@clr/ui": "^4.0.3",
     "@ng-select/ng-select": "^3.7.2",
     "@ngx-translate/core": "^12.1.2",
     "@ngx-translate/http-loader": "^4.0.0",
@@ -61,15 +61,15 @@
     "prosemirror-schema-list": "^1.0.0",
     "prosemirror-state": "^1.0.0",
     "rxjs": "^6.5.4",
-    "tslib": "^1.10.0",
+    "tslib": "^2.0.0",
     "zone.js": "~0.10.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.901.6",
-    "@angular-devkit/build-ng-packagr": "~0.901.6",
-    "@angular/cli": "^9.1.6",
-    "@angular/compiler": "^9.1.7",
-    "@angular/compiler-cli": "^9.1.7",
+    "@angular-devkit/build-angular": "~0.1001.4",
+    "@angular-devkit/build-ng-packagr": "~0.1001.4",
+    "@angular/cli": "^10.1.4",
+    "@angular/compiler": "^10.1.4",
+    "@angular/compiler-cli": "^10.1.4",
     "@biesbjerg/ngx-translate-extract": "^6.0.3",
     "@types/jasmine": "~3.5.10",
     "@types/jasminewd2": "~2.0.6",
@@ -82,18 +82,18 @@
     "cross-spawn": "^7.0.3",
     "fs-extra": "^9.0.0",
     "jasmine-core": "~3.5.0",
-    "jasmine-spec-reporter": "~5.0.1",
-    "karma": "~4.4.1",
+    "jasmine-spec-reporter": "~5.0.0",
+    "karma": "~5.0.0",
     "karma-chrome-launcher": "~3.1.0",
-    "karma-coverage-istanbul-reporter": "~2.1.1",
-    "karma-jasmine": "~3.1.1",
-    "karma-jasmine-html-reporter": "^1.5.3",
+    "karma-coverage-istanbul-reporter": "~3.0.2",
+    "karma-jasmine": "~4.0.0",
+    "karma-jasmine-html-reporter": "^1.5.0",
     "karma-mocha-reporter": "^2.2.5",
-    "ng-packagr": "^9.0.0",
-    "protractor": "~5.4.2",
+    "ng-packagr": "^10.1.0",
+    "protractor": "~7.0.0",
     "puppeteer": "^2.1.1",
     "rimraf": "^3.0.0",
-    "tslint": "^6.1.0",
-    "typescript": "3.8.3"
+    "tslint": "~6.1.0",
+    "typescript": "4.0.3"
   }
 }
