@@ -2647,6 +2647,7 @@ export type Order = Node & {
   promotions: Array<Promotion>;
   payments?: Maybe<Array<Payment>>;
   fulfillments?: Maybe<Array<Fulfillment>>;
+  totalQuantity: Scalars['Int'];
   subTotalBeforeTax: Scalars['Int'];
   /** The subTotal is the total of the OrderLines, before order-level promotions and shipping has been applied. */
   subTotal: Scalars['Int'];
@@ -2685,6 +2686,7 @@ export type OrderFilterParameter = {
   code?: Maybe<StringOperators>;
   state?: Maybe<StringOperators>;
   active?: Maybe<BooleanOperators>;
+  totalQuantity?: Maybe<NumberOperators>;
   subTotalBeforeTax?: Maybe<NumberOperators>;
   subTotal?: Maybe<NumberOperators>;
   currencyCode?: Maybe<StringOperators>;
@@ -2756,6 +2758,7 @@ export type OrderSortParameter = {
   updatedAt?: Maybe<SortOrder>;
   code?: Maybe<SortOrder>;
   state?: Maybe<SortOrder>;
+  totalQuantity?: Maybe<SortOrder>;
   subTotalBeforeTax?: Maybe<SortOrder>;
   subTotal?: Maybe<SortOrder>;
   shipping?: Maybe<SortOrder>;
@@ -6906,7 +6909,7 @@ export type TestShippingMethodQuery = { testShippingMethod: (
     & Pick<TestShippingMethodResult, 'eligible'>
     & { quote?: Maybe<(
       { __typename?: 'TestShippingMethodQuote' }
-      & Pick<TestShippingMethodQuote, 'price' | 'priceWithTax' | 'metadata'>
+      & Pick<TestShippingMethodQuote, 'price' | 'priceWithTax' | 'description' | 'metadata'>
     )> }
   ) };
 
